refactor: move genMonthMatrix to separated file
