+++ src/components/CalendarPicker/index.js
@@ -7,31 +7,7 @@ import ListPicker from './ListPicker';
 import ArrowIcon from './ArrowIcon';
 import styles from '../../styles/styles';
 import {propTypes, defaultProps} from './calendarPickerPropTypes';
-
-function generateMonthMatrix(year, month) {
-    const daysInMonth = moment([year, month]).daysInMonth();
-    const firstDay = moment([year, month, 1]).startOf('month').locale('en');
-    const matrix = [];
-    let currentWeek = [];
-    for (let i = 0; i < firstDay.weekday(); i++) {
-        currentWeek.push(null);
-    }
-    for (let i = 1; i <= daysInMonth; i++) {
-        const day = moment([year, month, i]).locale('en');
-        currentWeek.push(day.date());
-        if (day.weekday() === 6) {
-            matrix.push(currentWeek);
-            currentWeek = [];
-        }
-    }
-    if (currentWeek.length > 0) {
-        for (let i = currentWeek.length; i < 7; i++) {
-            currentWeek.push(null);
-        }
-        matrix.push(currentWeek);
-    }
-    return matrix;
-}
+import generateMonthMatrix from './generateMonthMatrix';
 
 const monthNames = ['January', 'February', 'March', 'April', 'May', 'June', 'July', 'August', 'September', 'October', 'November', 'December'];
 const daysOfWeek = ['Sunday', 'Monday', 'Tuesday', 'Wednesday', 'Thursday', 'Friday', 'Saturday'];
